feat: hide stack trace behind `Collapse`
